+++ src/main/kotlin/com/github/fantom/codeowners/lang/kind/bitbucket/psi/impl/CodeownersTeamDefinitionExtImpl.kt
@@ -1,0 +1,17 @@
+package com.github.fantom.codeowners.lang.kind.bitbucket.psi.impl
+
+import com.github.fantom.codeowners.lang.kind.bitbucket.psi.*
+import com.intellij.lang.ASTNode
+import com.intellij.navigation.ItemPresentation
+import javax.swing.Icon
+
+abstract class CodeownersTeamDefinitionExtImpl(node: ASTNode) : CodeownersElementImpl(node), CodeownersTeamDefinition {
+    override fun getName(): String = teamName.text
+
+    override fun getPresentation(): ItemPresentation? {
+        return object : ItemPresentation {
+            override fun getPresentableText(): String = name
+            override fun getIcon(unused: Boolean): Icon? = null
+        }
+    }
+}