+++ src/main/kotlin/com/github/fantom/codeowners/structureview/StructureViewFactory.kt
@@ -1,0 +1,18 @@
+package com.github.fantom.codeowners.structureview
+
+import com.intellij.ide.structureView.StructureViewBuilder
+import com.intellij.ide.structureView.StructureViewModel
+import com.intellij.ide.structureView.TreeBasedStructureViewBuilder
+import com.intellij.lang.PsiStructureViewFactory
+import com.intellij.openapi.editor.Editor
+import com.intellij.psi.PsiFile
+
+class StructureViewFactory : PsiStructureViewFactory {
+    override fun getStructureViewBuilder(psiFile: PsiFile): StructureViewBuilder {
+        return object : TreeBasedStructureViewBuilder() {
+            override fun createStructureViewModel(editor: Editor?): StructureViewModel {
+                return StructureViewModel(psiFile)
+            }
+        }
+    }
+}